@@ -1,26 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="coverlet.msbuild" Version="2.7.0">
+        <PackageReference Include="coverlet.msbuild" Version="2.8.0">
           <PrivateAssets>all</PrivateAssets>
           <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="FluentAssertions" Version="5.9.0" />
-        <PackageReference Include="Microsoft.TestPlatform" Version="16.3.0" />
+        <PackageReference Include="FluentAssertions" Version="5.10.0" />
+        <PackageReference Include="Microsoft.TestPlatform" Version="16.4.0" />
         <PackageReference Include="Moq" Version="4.13.1" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.3.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
         <PackageReference Include="NUnit" Version="3.12.0" />
-        <PackageReference Include="dotnet-reportgenerator-cli" Version="4.1.7" />
-        <PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
-        <PackageReference Include="Microsoft.AspNetCore.App" />
-        <PackageReference Include="NBuilder" Version="6.0.1" />
-    </ItemGroup>
-    <ItemGroup>
-        <DotNetCliToolReference Include="dotnet-reportgenerator-cli" Version="4.0.11" />
+        <PackageReference Include="NUnit3TestAdapter" Version="3.16.1" />
+        <PackageReference Include="NBuilder" Version="6.1.0" />
     </ItemGroup>
     <ItemGroup>
       <ProjectReference Include="..\Testing.Common\Testing.Common.csproj" />
