@@ -2,7 +2,7 @@
 /* eslint-disable */
 //----------------------
 // <auto-generated>
-//     Generated using the NSwag toolchain v13.0.6.0 (NJsonSchema v10.0.23.0 (Newtonsoft.Json v11.0.0.0)) (http://NSwag.org)
+//     Generated using the NSwag toolchain v13.1.3.0 (NJsonSchema v10.0.27.0 (Newtonsoft.Json v11.0.0.0)) (http://NSwag.org)
 // </auto-generated>
 //----------------------
 // ReSharper disable InconsistentNaming
@@ -342,6 +342,10 @@ export class ApiClient {
             result200 = ClientSettingsResponse.fromJS(resultData200);
             return _observableOf(result200);
             }));
+        } else if (status === 401) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("Unauthorized", status, _responseText, _headers);
+            }));
         } else if (status !== 200 && status !== 204) {
             return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
             return throwException("An unexpected server error occurred.", status, _responseText, _headers);
@@ -352,21 +356,21 @@ export class ApiClient {
 
     /**
      * Raise or answer to a private consultation request with another participant
-     * @param request (optional) Private consultation request with or without an answer
+     * @param body (optional) Private consultation request with or without an answer
      * @return Success
      */
-    handleConsultationRequest(request: ConsultationRequest | null | undefined): Observable<void> {
+    handleConsultationRequest(body: ConsultationRequest | undefined): Observable<void> {
         let url_ = this.baseUrl + "/consultations";
         url_ = url_.replace(/[?&]$/, "");
 
-        const content_ = JSON.stringify(request);
+        const content_ = JSON.stringify(body);
 
         let options_ : any = {
             body: content_,
             observe: "response",
             responseType: "blob",
             headers: new HttpHeaders({
-                "Content-Type": "application/json", 
+                "Content-Type": "application/json-patch+json", 
             })
         };
 
@@ -415,21 +419,21 @@ export class ApiClient {
     }
 
     /**
-     * @param request (optional) 
+     * @param body (optional) 
      * @return Success
      */
-    leavePrivateConsultation(request: LeaveConsultationRequest | null | undefined): Observable<void> {
+    leavePrivateConsultation(body: LeaveConsultationRequest | undefined): Observable<void> {
         let url_ = this.baseUrl + "/consultations/leave";
         url_ = url_.replace(/[?&]$/, "");
 
-        const content_ = JSON.stringify(request);
+        const content_ = JSON.stringify(body);
 
         let options_ : any = {
             body: content_,
             observe: "response",
             responseType: "blob",
             headers: new HttpHeaders({
-                "Content-Type": "application/json", 
+                "Content-Type": "application/json-patch+json", 
             })
         };
 
@@ -485,21 +489,21 @@ export class ApiClient {
     }
 
     /**
-     * @param request (optional) 
+     * @param body (optional) 
      * @return Success
      */
-    respondToAdminConsultationRequest(request: AdminConsultationRequest | null | undefined): Observable<void> {
+    respondToAdminConsultationRequest(body: AdminConsultationRequest | undefined): Observable<void> {
         let url_ = this.baseUrl + "/consultations/vhofficer/respond";
         url_ = url_.replace(/[?&]$/, "");
 
-        const content_ = JSON.stringify(request);
+        const content_ = JSON.stringify(body);
 
         let options_ : any = {
             body: content_,
             observe: "response",
             responseType: "blob",
             headers: new HttpHeaders({
-                "Content-Type": "application/json", 
+                "Content-Type": "application/json-patch+json", 
             })
         };
 
@@ -605,6 +609,10 @@ export class ApiClient {
             result500 = HealthCheckResponse.fromJS(resultData500);
             return throwException("Server Error", status, _responseText, _headers, result500);
             }));
+        } else if (status === 401) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("Unauthorized", status, _responseText, _headers);
+            }));
         } else if (status !== 200 && status !== 204) {
             return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
             return throwException("An unexpected server error occurred.", status, _responseText, _headers);
@@ -614,24 +622,24 @@ export class ApiClient {
     }
 
     /**
-     * @param addMediaEventRequest (optional) 
+     * @param body (optional) 
      * @return Success
      */
-    addMediaEventToConference(conferenceId: string, addMediaEventRequest: AddMediaEventRequest | null | undefined): Observable<void> {
+    addMediaEventToConference(conferenceId: string, body: AddMediaEventRequest | undefined): Observable<void> {
         let url_ = this.baseUrl + "/conferences/{conferenceId}/mediaevents";
         if (conferenceId === undefined || conferenceId === null)
             throw new Error("The parameter 'conferenceId' must be defined.");
         url_ = url_.replace("{conferenceId}", encodeURIComponent("" + conferenceId)); 
         url_ = url_.replace(/[?&]$/, "");
 
-        const content_ = JSON.stringify(addMediaEventRequest);
+        const content_ = JSON.stringify(body);
 
         let options_ : any = {
             body: content_,
             observe: "response",
             responseType: "blob",
             headers: new HttpHeaders({
-                "Content-Type": "application/json", 
+                "Content-Type": "application/json-patch+json", 
             })
         };
 
@@ -687,24 +695,24 @@ export class ApiClient {
     }
 
     /**
-     * @param addSelfTestFailureEventRequest (optional) 
+     * @param body (optional) 
      * @return Success
      */
-    addSelfTestFailureEventToConference(conferenceId: string, addSelfTestFailureEventRequest: AddSelfTestFailureEventRequest | null | undefined): Observable<void> {
+    addSelfTestFailureEventToConference(conferenceId: string, body: AddSelfTestFailureEventRequest | undefined): Observable<void> {
         let url_ = this.baseUrl + "/conferences/{conferenceId}/selftestfailureevents";
         if (conferenceId === undefined || conferenceId === null)
             throw new Error("The parameter 'conferenceId' must be defined.");
         url_ = url_.replace("{conferenceId}", encodeURIComponent("" + conferenceId)); 
         url_ = url_.replace(/[?&]$/, "");
 
-        const content_ = JSON.stringify(addSelfTestFailureEventRequest);
+        const content_ = JSON.stringify(body);
 
         let options_ : any = {
             body: content_,
             observe: "response",
             responseType: "blob",
             headers: new HttpHeaders({
-                "Content-Type": "application/json", 
+                "Content-Type": "application/json-patch+json", 
             })
         };
 
@@ -828,24 +836,24 @@ export class ApiClient {
     }
 
     /**
-     * @param updateParticipantStatusEventRequest (optional) 
+     * @param body (optional) 
      * @return Success
      */
-    updateParticipantStatus(conferenceId: string, updateParticipantStatusEventRequest: UpdateParticipantStatusEventRequest | null | undefined): Observable<void> {
+    updateParticipantStatus(conferenceId: string, body: UpdateParticipantStatusEventRequest | undefined): Observable<void> {
         let url_ = this.baseUrl + "/conferences/{conferenceId}/participantstatus";
         if (conferenceId === undefined || conferenceId === null)
             throw new Error("The parameter 'conferenceId' must be defined.");
         url_ = url_.replace("{conferenceId}", encodeURIComponent("" + conferenceId)); 
         url_ = url_.replace(/[?&]$/, "");
 
-        const content_ = JSON.stringify(updateParticipantStatusEventRequest);
+        const content_ = JSON.stringify(body);
 
         let options_ : any = {
             body: content_,
             observe: "response",
             responseType: "blob",
             headers: new HttpHeaders({
-                "Content-Type": "application/json", 
+                "Content-Type": "application/json-patch+json", 
             })
         };
 
@@ -904,9 +912,11 @@ export class ApiClient {
      * @param participantId (optional) 
      * @return Success
      */
-    getIndependentTestCallResult(participantId: string | null | undefined): Observable<TestCallScoreResponse> {
+    getIndependentTestCallResult(participantId: string | undefined): Observable<TestCallScoreResponse> {
         let url_ = this.baseUrl + "/conferences/independentselftestresult?";
-        if (participantId !== undefined)
+        if (participantId === null)
+            throw new Error("The parameter 'participantId' cannot be null.");
+        else if (participantId !== undefined)
             url_ += "participantId=" + encodeURIComponent("" + participantId) + "&"; 
         url_ = url_.replace(/[?&]$/, "");
 
@@ -1354,21 +1364,88 @@ export class ApiClient {
     }
 
     /**
-     * @param request (optional) 
+     * Get hearings venues
      * @return Success
      */
-    sendEvent(request: ConferenceEventRequest | null | undefined): Observable<void> {
+    getHearingsVenues(): Observable<HearingVenueResponse[]> {
+        let url_ = this.baseUrl + "/hearings-venues";
+        url_ = url_.replace(/[?&]$/, "");
+
+        let options_ : any = {
+            observe: "response",
+            responseType: "blob",
+            headers: new HttpHeaders({
+                "Accept": "application/json"
+            })
+        };
+
+        return this.http.request("get", url_, options_).pipe(_observableMergeMap((response_ : any) => {
+            return this.processGetHearingsVenues(response_);
+        })).pipe(_observableCatch((response_: any) => {
+            if (response_ instanceof HttpResponseBase) {
+                try {
+                    return this.processGetHearingsVenues(<any>response_);
+                } catch (e) {
+                    return <Observable<HearingVenueResponse[]>><any>_observableThrow(e);
+                }
+            } else
+                return <Observable<HearingVenueResponse[]>><any>_observableThrow(response_);
+        }));
+    }
+
+    protected processGetHearingsVenues(response: HttpResponseBase): Observable<HearingVenueResponse[]> {
+        const status = response.status;
+        const responseBlob = 
+            response instanceof HttpResponse ? response.body : 
+            (<any>response).error instanceof Blob ? (<any>response).error : undefined;
+
+        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }};
+        if (status === 200) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result200: any = null;
+            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            if (Array.isArray(resultData200)) {
+                result200 = [] as any;
+                for (let item of resultData200)
+                    result200!.push(HearingVenueResponse.fromJS(item));
+            }
+            return _observableOf(result200);
+            }));
+        } else if (status === 400) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            let result400: any = null;
+            let resultData400 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result400 = ProblemDetails.fromJS(resultData400);
+            return throwException("Bad Request", status, _responseText, _headers, result400);
+            }));
+        } else if (status === 401) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("Unauthorized", status, _responseText, _headers);
+            }));
+        } else if (status !== 200 && status !== 204) {
+            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
+            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
+            }));
+        }
+        return _observableOf<HearingVenueResponse[]>(<any>null);
+    }
+
+    /**
+     * @param body (optional) 
+     * @return Success
+     */
+    sendEvent(body: ConferenceEventRequest | undefined): Observable<void> {
         let url_ = this.baseUrl + "/callback";
         url_ = url_.replace(/[?&]$/, "");
 
-        const content_ = JSON.stringify(request);
+        const content_ = JSON.stringify(body);
 
         let options_ : any = {
             body: content_,
             observe: "response",
             responseType: "blob",
             headers: new HttpHeaders({
-                "Content-Type": "application/json", 
+                "Content-Type": "application/json-patch+json", 
             })
         };
 
@@ -1417,19 +1494,162 @@ export class ApiClient {
     }
 }
 
+/** Known states of a conference */
+export enum ConferenceStatus {
+    NotStarted = "NotStarted",
+    InSession = "InSession",
+    Paused = "Paused",
+    Suspended = "Suspended",
+    Closed = "Closed",
+}
+
+/** Known user roles */
+export enum UserRole {
+    None = "None",
+    CaseAdmin = "CaseAdmin",
+    VideoHearingsOfficer = "VideoHearingsOfficer",
+    HearingFacilitationSupport = "HearingFacilitationSupport",
+    Judge = "Judge",
+    Individual = "Individual",
+    Representative = "Representative",
+}
+
+/** Known states of a participant */
+export enum ParticipantStatus {
+    None = "None",
+    NotSignedIn = "NotSignedIn",
+    UnableToJoin = "UnableToJoin",
+    Joining = "Joining",
+    Available = "Available",
+    InHearing = "InHearing",
+    InConsultation = "InConsultation",
+    Disconnected = "Disconnected",
+}
+
+export class ParticipantForUserResponse implements IParticipantForUserResponse {
+    /** The participant username */
+    username?: string | undefined;
+    display_name?: string | undefined;
+    /** The participant role in conference */
+    role?: UserRole;
+    /** The current status of a participant */
+    status?: ParticipantStatus;
+    /** The representee (if participant is a representative) */
+    representee?: string | undefined;
+    case_type_group?: string | undefined;
+
+    constructor(data?: IParticipantForUserResponse) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.username = _data["username"];
+            this.display_name = _data["display_name"];
+            this.role = _data["role"];
+            this.status = _data["status"];
+            this.representee = _data["representee"];
+            this.case_type_group = _data["case_type_group"];
+        }
+    }
+
+    static fromJS(data: any): ParticipantForUserResponse {
+        data = typeof data === 'object' ? data : {};
+        let result = new ParticipantForUserResponse();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["username"] = this.username;
+        data["display_name"] = this.display_name;
+        data["role"] = this.role;
+        data["status"] = this.status;
+        data["representee"] = this.representee;
+        data["case_type_group"] = this.case_type_group;
+        return data; 
+    }
+}
+
+export interface IParticipantForUserResponse {
+    /** The participant username */
+    username?: string | undefined;
+    display_name?: string | undefined;
+    /** The participant role in conference */
+    role?: UserRole;
+    /** The current status of a participant */
+    status?: ParticipantStatus;
+    /** The representee (if participant is a representative) */
+    representee?: string | undefined;
+    case_type_group?: string | undefined;
+}
+
+export class TaskUserResponse implements ITaskUserResponse {
+    id?: number;
+    body?: string | undefined;
+
+    constructor(data?: ITaskUserResponse) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.body = _data["body"];
+        }
+    }
+
+    static fromJS(data: any): TaskUserResponse {
+        data = typeof data === 'object' ? data : {};
+        let result = new TaskUserResponse();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["id"] = this.id;
+        data["body"] = this.body;
+        return data; 
+    }
+}
+
+export interface ITaskUserResponse {
+    id?: number;
+    body?: string | undefined;
+}
+
+/** High level summary of a conference for a user */
 export class ConferenceForUserResponse implements IConferenceForUserResponse {
-    id?: string | undefined;
-    scheduled_date_time?: Date | undefined;
+    /** Conference ID */
+    id?: string;
+    scheduled_date_time?: Date;
     case_type?: string | undefined;
     case_number?: string | undefined;
     case_name?: string | undefined;
-    scheduled_duration?: number | undefined;
-    status?: ConferenceStatus | undefined;
+    scheduled_duration?: number;
+    /** The current conference status */
+    status?: ConferenceStatus;
+    /** The conference participants */
     participants?: ParticipantForUserResponse[] | undefined;
-    no_of_participants_available?: number | undefined;
-    no_of_participants_unavailable?: number | undefined;
-    no_of_participants_in_consultation?: number | undefined;
-    no_of_pending_tasks?: number | undefined;
+    no_of_participants_available?: number;
+    no_of_participants_unavailable?: number;
+    no_of_participants_in_consultation?: number;
+    no_of_pending_tasks?: number;
+    hearing_venue_name?: string | undefined;
+    /** The conferences tasks */
+    tasks?: TaskUserResponse[] | undefined;
 
     constructor(data?: IConferenceForUserResponse) {
         if (data) {
@@ -1440,24 +1660,30 @@ export class ConferenceForUserResponse implements IConferenceForUserResponse {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.id = data["id"];
-            this.scheduled_date_time = data["scheduled_date_time"] ? new Date(data["scheduled_date_time"].toString()) : <any>undefined;
-            this.case_type = data["case_type"];
-            this.case_number = data["case_number"];
-            this.case_name = data["case_name"];
-            this.scheduled_duration = data["scheduled_duration"];
-            this.status = data["status"];
-            if (Array.isArray(data["participants"])) {
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.scheduled_date_time = _data["scheduled_date_time"] ? new Date(_data["scheduled_date_time"].toString()) : <any>undefined;
+            this.case_type = _data["case_type"];
+            this.case_number = _data["case_number"];
+            this.case_name = _data["case_name"];
+            this.scheduled_duration = _data["scheduled_duration"];
+            this.status = _data["status"];
+            if (Array.isArray(_data["participants"])) {
                 this.participants = [] as any;
-                for (let item of data["participants"])
+                for (let item of _data["participants"])
                     this.participants!.push(ParticipantForUserResponse.fromJS(item));
             }
-            this.no_of_participants_available = data["no_of_participants_available"];
-            this.no_of_participants_unavailable = data["no_of_participants_unavailable"];
-            this.no_of_participants_in_consultation = data["no_of_participants_in_consultation"];
-            this.no_of_pending_tasks = data["no_of_pending_tasks"];
+            this.no_of_participants_available = _data["no_of_participants_available"];
+            this.no_of_participants_unavailable = _data["no_of_participants_unavailable"];
+            this.no_of_participants_in_consultation = _data["no_of_participants_in_consultation"];
+            this.no_of_pending_tasks = _data["no_of_pending_tasks"];
+            this.hearing_venue_name = _data["hearing_venue_name"];
+            if (Array.isArray(_data["tasks"])) {
+                this.tasks = [] as any;
+                for (let item of _data["tasks"])
+                    this.tasks!.push(TaskUserResponse.fromJS(item));
+            }
         }
     }
 
@@ -1486,108 +1712,36 @@ export class ConferenceForUserResponse implements IConferenceForUserResponse {
         data["no_of_participants_unavailable"] = this.no_of_participants_unavailable;
         data["no_of_participants_in_consultation"] = this.no_of_participants_in_consultation;
         data["no_of_pending_tasks"] = this.no_of_pending_tasks;
+        data["hearing_venue_name"] = this.hearing_venue_name;
+        if (Array.isArray(this.tasks)) {
+            data["tasks"] = [];
+            for (let item of this.tasks)
+                data["tasks"].push(item.toJSON());
+        }
         return data; 
     }
 }
 
+/** High level summary of a conference for a user */
 export interface IConferenceForUserResponse {
-    id?: string | undefined;
-    scheduled_date_time?: Date | undefined;
+    /** Conference ID */
+    id?: string;
+    scheduled_date_time?: Date;
     case_type?: string | undefined;
     case_number?: string | undefined;
     case_name?: string | undefined;
-    scheduled_duration?: number | undefined;
-    status?: ConferenceStatus | undefined;
+    scheduled_duration?: number;
+    /** The current conference status */
+    status?: ConferenceStatus;
+    /** The conference participants */
     participants?: ParticipantForUserResponse[] | undefined;
-    no_of_participants_available?: number | undefined;
-    no_of_participants_unavailable?: number | undefined;
-    no_of_participants_in_consultation?: number | undefined;
-    no_of_pending_tasks?: number | undefined;
-}
-
-export enum ConferenceStatus {
-    NotStarted = "NotStarted",
-    InSession = "InSession",
-    Paused = "Paused",
-    Suspended = "Suspended",
-    Closed = "Closed",
-}
-
-export class ParticipantForUserResponse implements IParticipantForUserResponse {
-    username?: string | undefined;
-    display_name?: string | undefined;
-    role?: UserRole | undefined;
-    status?: ParticipantStatus | undefined;
-    representee?: string | undefined;
-    case_type_group?: string | undefined;
-
-    constructor(data?: IParticipantForUserResponse) {
-        if (data) {
-            for (var property in data) {
-                if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
-            }
-        }
-    }
-
-    init(data?: any) {
-        if (data) {
-            this.username = data["username"];
-            this.display_name = data["display_name"];
-            this.role = data["role"];
-            this.status = data["status"];
-            this.representee = data["representee"];
-            this.case_type_group = data["case_type_group"];
-        }
-    }
-
-    static fromJS(data: any): ParticipantForUserResponse {
-        data = typeof data === 'object' ? data : {};
-        let result = new ParticipantForUserResponse();
-        result.init(data);
-        return result;
-    }
-
-    toJSON(data?: any) {
-        data = typeof data === 'object' ? data : {};
-        data["username"] = this.username;
-        data["display_name"] = this.display_name;
-        data["role"] = this.role;
-        data["status"] = this.status;
-        data["representee"] = this.representee;
-        data["case_type_group"] = this.case_type_group;
-        return data; 
-    }
-}
-
-export interface IParticipantForUserResponse {
-    username?: string | undefined;
-    display_name?: string | undefined;
-    role?: UserRole | undefined;
-    status?: ParticipantStatus | undefined;
-    representee?: string | undefined;
-    case_type_group?: string | undefined;
-}
-
-export enum UserRole {
-    None = "None",
-    CaseAdmin = "CaseAdmin",
-    VideoHearingsOfficer = "VideoHearingsOfficer",
-    HearingFacilitationSupport = "HearingFacilitationSupport",
-    Judge = "Judge",
-    Individual = "Individual",
-    Representative = "Representative",
-}
-
-export enum ParticipantStatus {
-    None = "None",
-    NotSignedIn = "NotSignedIn",
-    UnableToJoin = "UnableToJoin",
-    Joining = "Joining",
-    Available = "Available",
-    InHearing = "InHearing",
-    InConsultation = "InConsultation",
-    Disconnected = "Disconnected",
+    no_of_participants_available?: number;
+    no_of_participants_unavailable?: number;
+    no_of_participants_in_consultation?: number;
+    no_of_pending_tasks?: number;
+    hearing_venue_name?: string | undefined;
+    /** The conferences tasks */
+    tasks?: TaskUserResponse[] | undefined;
 }
 
 export class ProblemDetails implements IProblemDetails {
@@ -1596,6 +1750,7 @@ export class ProblemDetails implements IProblemDetails {
     status?: number | undefined;
     detail?: string | undefined;
     instance?: string | undefined;
+    readonly extensions?: { [key: string]: any; } | undefined;
 
     constructor(data?: IProblemDetails) {
         if (data) {
@@ -1606,13 +1761,20 @@ export class ProblemDetails implements IProblemDetails {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.type = data["type"];
-            this.title = data["title"];
-            this.status = data["status"];
-            this.detail = data["detail"];
-            this.instance = data["instance"];
+    init(_data?: any) {
+        if (_data) {
+            this.type = _data["type"];
+            this.title = _data["title"];
+            this.status = _data["status"];
+            this.detail = _data["detail"];
+            this.instance = _data["instance"];
+            if (_data["extensions"]) {
+                (<any>this).extensions = {} as any;
+                for (let key in _data["extensions"]) {
+                    if (_data["extensions"].hasOwnProperty(key))
+                        (<any>this).extensions![key] = _data["extensions"][key];
+                }
+            }
         }
     }
 
@@ -1630,6 +1792,13 @@ export class ProblemDetails implements IProblemDetails {
         data["status"] = this.status;
         data["detail"] = this.detail;
         data["instance"] = this.instance;
+        if (this.extensions) {
+            data["extensions"] = {};
+            for (let key in this.extensions) {
+                if (this.extensions.hasOwnProperty(key))
+                    data["extensions"][key] = this.extensions[key];
+            }
+        }
         return data; 
     }
 }
@@ -1640,21 +1809,124 @@ export interface IProblemDetails {
     status?: number | undefined;
     detail?: string | undefined;
     instance?: string | undefined;
+    extensions?: { [key: string]: any; } | undefined;
 }
 
+/** Information about a participant in a conference */
+export class ParticipantResponse implements IParticipantResponse {
+    /** The participant id in a conference */
+    id?: string;
+    first_name?: string | undefined;
+    last_name?: string | undefined;
+    contact_email?: string | undefined;
+    contact_telephone?: string | undefined;
+    /** The participant's full name */
+    name?: string | undefined;
+    /** The participant's username */
+    username?: string | undefined;
+    /** The participant's role */
+    role?: UserRole;
+    /** The participant's status */
+    status?: ParticipantStatus;
+    display_name?: string | undefined;
+    tiled_display_name?: string | undefined;
+    case_type_group?: string | undefined;
+    /** The representee the participant is acting on behalf */
+    representee?: string | undefined;
+
+    constructor(data?: IParticipantResponse) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.first_name = _data["first_name"];
+            this.last_name = _data["last_name"];
+            this.contact_email = _data["contact_email"];
+            this.contact_telephone = _data["contact_telephone"];
+            this.name = _data["name"];
+            this.username = _data["username"];
+            this.role = _data["role"];
+            this.status = _data["status"];
+            this.display_name = _data["display_name"];
+            this.tiled_display_name = _data["tiled_display_name"];
+            this.case_type_group = _data["case_type_group"];
+            this.representee = _data["representee"];
+        }
+    }
+
+    static fromJS(data: any): ParticipantResponse {
+        data = typeof data === 'object' ? data : {};
+        let result = new ParticipantResponse();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["id"] = this.id;
+        data["first_name"] = this.first_name;
+        data["last_name"] = this.last_name;
+        data["contact_email"] = this.contact_email;
+        data["contact_telephone"] = this.contact_telephone;
+        data["name"] = this.name;
+        data["username"] = this.username;
+        data["role"] = this.role;
+        data["status"] = this.status;
+        data["display_name"] = this.display_name;
+        data["tiled_display_name"] = this.tiled_display_name;
+        data["case_type_group"] = this.case_type_group;
+        data["representee"] = this.representee;
+        return data; 
+    }
+}
+
+/** Information about a participant in a conference */
+export interface IParticipantResponse {
+    /** The participant id in a conference */
+    id?: string;
+    first_name?: string | undefined;
+    last_name?: string | undefined;
+    contact_email?: string | undefined;
+    contact_telephone?: string | undefined;
+    /** The participant's full name */
+    name?: string | undefined;
+    /** The participant's username */
+    username?: string | undefined;
+    /** The participant's role */
+    role?: UserRole;
+    /** The participant's status */
+    status?: ParticipantStatus;
+    display_name?: string | undefined;
+    tiled_display_name?: string | undefined;
+    case_type_group?: string | undefined;
+    /** The representee the participant is acting on behalf */
+    representee?: string | undefined;
+}
+
+/** Detailed information about a conference */
 export class ConferenceResponse implements IConferenceResponse {
-    id?: string | undefined;
-    scheduled_date_time?: Date | undefined;
-    scheduled_duration?: number | undefined;
+    /** Conference ID */
+    id?: string;
+    scheduled_date_time?: Date;
+    scheduled_duration?: number;
     case_type?: string | undefined;
     case_number?: string | undefined;
     case_name?: string | undefined;
-    status?: ConferenceStatus | undefined;
+    /** The current conference Status */
+    status?: ConferenceStatus;
     judge_i_frame_uri?: string | undefined;
     admin_i_frame_uri?: string | undefined;
     participant_uri?: string | undefined;
     pexip_node_uri?: string | undefined;
     pexip_self_test_node_uri?: string | undefined;
+    /** The participants in the conference */
     participants?: ParticipantResponse[] | undefined;
     closed_date_time?: Date | undefined;
 
@@ -1667,26 +1939,26 @@ export class ConferenceResponse implements IConferenceResponse {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.id = data["id"];
-            this.scheduled_date_time = data["scheduled_date_time"] ? new Date(data["scheduled_date_time"].toString()) : <any>undefined;
-            this.scheduled_duration = data["scheduled_duration"];
-            this.case_type = data["case_type"];
-            this.case_number = data["case_number"];
-            this.case_name = data["case_name"];
-            this.status = data["status"];
-            this.judge_i_frame_uri = data["judge_i_frame_uri"];
-            this.admin_i_frame_uri = data["admin_i_frame_uri"];
-            this.participant_uri = data["participant_uri"];
-            this.pexip_node_uri = data["pexip_node_uri"];
-            this.pexip_self_test_node_uri = data["pexip_self_test_node_uri"];
-            if (Array.isArray(data["participants"])) {
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.scheduled_date_time = _data["scheduled_date_time"] ? new Date(_data["scheduled_date_time"].toString()) : <any>undefined;
+            this.scheduled_duration = _data["scheduled_duration"];
+            this.case_type = _data["case_type"];
+            this.case_number = _data["case_number"];
+            this.case_name = _data["case_name"];
+            this.status = _data["status"];
+            this.judge_i_frame_uri = _data["judge_i_frame_uri"];
+            this.admin_i_frame_uri = _data["admin_i_frame_uri"];
+            this.participant_uri = _data["participant_uri"];
+            this.pexip_node_uri = _data["pexip_node_uri"];
+            this.pexip_self_test_node_uri = _data["pexip_self_test_node_uri"];
+            if (Array.isArray(_data["participants"])) {
                 this.participants = [] as any;
-                for (let item of data["participants"])
+                for (let item of _data["participants"])
                     this.participants!.push(ParticipantResponse.fromJS(item));
             }
-            this.closed_date_time = data["closed_date_time"] ? new Date(data["closed_date_time"].toString()) : <any>undefined;
+            this.closed_date_time = _data["closed_date_time"] ? new Date(_data["closed_date_time"].toString()) : <any>undefined;
         }
     }
 
@@ -1721,107 +1993,28 @@ export class ConferenceResponse implements IConferenceResponse {
     }
 }
 
+/** Detailed information about a conference */
 export interface IConferenceResponse {
-    id?: string | undefined;
-    scheduled_date_time?: Date | undefined;
-    scheduled_duration?: number | undefined;
+    /** Conference ID */
+    id?: string;
+    scheduled_date_time?: Date;
+    scheduled_duration?: number;
     case_type?: string | undefined;
     case_number?: string | undefined;
     case_name?: string | undefined;
-    status?: ConferenceStatus | undefined;
+    /** The current conference Status */
+    status?: ConferenceStatus;
     judge_i_frame_uri?: string | undefined;
     admin_i_frame_uri?: string | undefined;
     participant_uri?: string | undefined;
     pexip_node_uri?: string | undefined;
     pexip_self_test_node_uri?: string | undefined;
+    /** The participants in the conference */
     participants?: ParticipantResponse[] | undefined;
     closed_date_time?: Date | undefined;
 }
 
-export class ParticipantResponse implements IParticipantResponse {
-    id?: string | undefined;
-    first_name?: string | undefined;
-    last_name?: string | undefined;
-    contact_email?: string | undefined;
-    contact_telephone?: string | undefined;
-    name?: string | undefined;
-    username?: string | undefined;
-    role?: UserRole | undefined;
-    status?: ParticipantStatus | undefined;
-    display_name?: string | undefined;
-    tiled_display_name?: string | undefined;
-    case_type_group?: string | undefined;
-    representee?: string | undefined;
-
-    constructor(data?: IParticipantResponse) {
-        if (data) {
-            for (var property in data) {
-                if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
-            }
-        }
-    }
-
-    init(data?: any) {
-        if (data) {
-            this.id = data["id"];
-            this.first_name = data["first_name"];
-            this.last_name = data["last_name"];
-            this.contact_email = data["contact_email"];
-            this.contact_telephone = data["contact_telephone"];
-            this.name = data["name"];
-            this.username = data["username"];
-            this.role = data["role"];
-            this.status = data["status"];
-            this.display_name = data["display_name"];
-            this.tiled_display_name = data["tiled_display_name"];
-            this.case_type_group = data["case_type_group"];
-            this.representee = data["representee"];
-        }
-    }
-
-    static fromJS(data: any): ParticipantResponse {
-        data = typeof data === 'object' ? data : {};
-        let result = new ParticipantResponse();
-        result.init(data);
-        return result;
-    }
-
-    toJSON(data?: any) {
-        data = typeof data === 'object' ? data : {};
-        data["id"] = this.id;
-        data["first_name"] = this.first_name;
-        data["last_name"] = this.last_name;
-        data["contact_email"] = this.contact_email;
-        data["contact_telephone"] = this.contact_telephone;
-        data["name"] = this.name;
-        data["username"] = this.username;
-        data["role"] = this.role;
-        data["status"] = this.status;
-        data["display_name"] = this.display_name;
-        data["tiled_display_name"] = this.tiled_display_name;
-        data["case_type_group"] = this.case_type_group;
-        data["representee"] = this.representee;
-        return data; 
-    }
-}
-
-export interface IParticipantResponse {
-    id?: string | undefined;
-    first_name?: string | undefined;
-    last_name?: string | undefined;
-    contact_email?: string | undefined;
-    contact_telephone?: string | undefined;
-    name?: string | undefined;
-    username?: string | undefined;
-    role?: UserRole | undefined;
-    status?: ParticipantStatus | undefined;
-    display_name?: string | undefined;
-    tiled_display_name?: string | undefined;
-    case_type_group?: string | undefined;
-    representee?: string | undefined;
-}
-
+/** Configuration to initialise the UI application */
 export class ClientSettingsResponse implements IClientSettingsResponse {
     tenant_id?: string | undefined;
     client_id?: string | undefined;
@@ -1839,14 +2032,14 @@ export class ClientSettingsResponse implements IClientSettingsResponse {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.tenant_id = data["tenant_id"];
-            this.client_id = data["client_id"];
-            this.redirect_uri = data["redirect_uri"];
-            this.post_logout_redirect_uri = data["post_logout_redirect_uri"];
-            this.video_api_url = data["video_api_url"];
-            this.app_insights_instrumentation_key = data["app_insights_instrumentation_key"];
+    init(_data?: any) {
+        if (_data) {
+            this.tenant_id = _data["tenant_id"];
+            this.client_id = _data["client_id"];
+            this.redirect_uri = _data["redirect_uri"];
+            this.post_logout_redirect_uri = _data["post_logout_redirect_uri"];
+            this.video_api_url = _data["video_api_url"];
+            this.app_insights_instrumentation_key = _data["app_insights_instrumentation_key"];
         }
     }
 
@@ -1869,6 +2062,7 @@ export class ClientSettingsResponse implements IClientSettingsResponse {
     }
 }
 
+/** Configuration to initialise the UI application */
 export interface IClientSettingsResponse {
     tenant_id?: string | undefined;
     client_id?: string | undefined;
@@ -1878,10 +2072,17 @@ export interface IClientSettingsResponse {
     app_insights_instrumentation_key?: string | undefined;
 }
 
+export enum ConsultationAnswer {
+    None = "None",
+    Accepted = "Accepted",
+    Rejected = "Rejected",
+    Cancelled = "Cancelled",
+}
+
 export class ConsultationRequest implements IConsultationRequest {
-    conference_id?: string | undefined;
-    requested_by?: string | undefined;
-    requested_for?: string | undefined;
+    conference_id!: string;
+    requested_by!: string;
+    requested_for!: string;
     answer?: ConsultationAnswer | undefined;
 
     constructor(data?: IConsultationRequest) {
@@ -1893,12 +2094,12 @@ export class ConsultationRequest implements IConsultationRequest {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.conference_id = data["conference_id"];
-            this.requested_by = data["requested_by"];
-            this.requested_for = data["requested_for"];
-            this.answer = data["answer"];
+    init(_data?: any) {
+        if (_data) {
+            this.conference_id = _data["conference_id"];
+            this.requested_by = _data["requested_by"];
+            this.requested_for = _data["requested_for"];
+            this.answer = _data["answer"];
         }
     }
 
@@ -1920,22 +2121,15 @@ export class ConsultationRequest implements IConsultationRequest {
 }
 
 export interface IConsultationRequest {
-    conference_id?: string | undefined;
-    requested_by?: string | undefined;
-    requested_for?: string | undefined;
+    conference_id: string;
+    requested_by: string;
+    requested_for: string;
     answer?: ConsultationAnswer | undefined;
 }
 
-export enum ConsultationAnswer {
-    None = "None",
-    Accepted = "Accepted",
-    Rejected = "Rejected",
-    Cancelled = "Cancelled",
-}
-
 export class LeaveConsultationRequest implements ILeaveConsultationRequest {
-    conference_id?: string | undefined;
-    participant_id?: string | undefined;
+    conference_id!: string;
+    participant_id!: string;
 
     constructor(data?: ILeaveConsultationRequest) {
         if (data) {
@@ -1946,10 +2140,10 @@ export class LeaveConsultationRequest implements ILeaveConsultationRequest {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.conference_id = data["conference_id"];
-            this.participant_id = data["participant_id"];
+    init(_data?: any) {
+        if (_data) {
+            this.conference_id = _data["conference_id"];
+            this.participant_id = _data["participant_id"];
         }
     }
 
@@ -1969,14 +2163,22 @@ export class LeaveConsultationRequest implements ILeaveConsultationRequest {
 }
 
 export interface ILeaveConsultationRequest {
-    conference_id?: string | undefined;
-    participant_id?: string | undefined;
+    conference_id: string;
+    participant_id: string;
+}
+
+export enum RoomType {
+    WaitingRoom = "WaitingRoom",
+    HearingRoom = "HearingRoom",
+    ConsultationRoom1 = "ConsultationRoom1",
+    ConsultationRoom2 = "ConsultationRoom2",
+    AdminRoom = "AdminRoom",
 }
 
 export class AdminConsultationRequest implements IAdminConsultationRequest {
-    conference_id?: string | undefined;
-    participant_id?: string | undefined;
-    consultation_room?: RoomType | undefined;
+    conference_id!: string;
+    participant_id!: string;
+    consultation_room?: RoomType;
     answer?: ConsultationAnswer | undefined;
 
     constructor(data?: IAdminConsultationRequest) {
@@ -1988,12 +2190,12 @@ export class AdminConsultationRequest implements IAdminConsultationRequest {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.conference_id = data["conference_id"];
-            this.participant_id = data["participant_id"];
-            this.consultation_room = data["consultation_room"];
-            this.answer = data["answer"];
+    init(_data?: any) {
+        if (_data) {
+            this.conference_id = _data["conference_id"];
+            this.participant_id = _data["participant_id"];
+            this.consultation_room = _data["consultation_room"];
+            this.answer = _data["answer"];
         }
     }
 
@@ -2015,66 +2217,14 @@ export class AdminConsultationRequest implements IAdminConsultationRequest {
 }
 
 export interface IAdminConsultationRequest {
-    conference_id?: string | undefined;
-    participant_id?: string | undefined;
-    consultation_room?: RoomType | undefined;
+    conference_id: string;
+    participant_id: string;
+    consultation_room?: RoomType;
     answer?: ConsultationAnswer | undefined;
 }
 
-export enum RoomType {
-    WaitingRoom = "WaitingRoom",
-    HearingRoom = "HearingRoom",
-    ConsultationRoom1 = "ConsultationRoom1",
-    ConsultationRoom2 = "ConsultationRoom2",
-    AdminRoom = "AdminRoom",
-}
-
-export class HealthCheckResponse implements IHealthCheckResponse {
-    bookings_api_health?: HealthCheck | undefined;
-    user_api_health?: HealthCheck | undefined;
-    video_api_health?: HealthCheck | undefined;
-
-    constructor(data?: IHealthCheckResponse) {
-        if (data) {
-            for (var property in data) {
-                if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
-            }
-        }
-    }
-
-    init(data?: any) {
-        if (data) {
-            this.bookings_api_health = data["bookings_api_health"] ? HealthCheck.fromJS(data["bookings_api_health"]) : <any>undefined;
-            this.user_api_health = data["user_api_health"] ? HealthCheck.fromJS(data["user_api_health"]) : <any>undefined;
-            this.video_api_health = data["video_api_health"] ? HealthCheck.fromJS(data["video_api_health"]) : <any>undefined;
-        }
-    }
-
-    static fromJS(data: any): HealthCheckResponse {
-        data = typeof data === 'object' ? data : {};
-        let result = new HealthCheckResponse();
-        result.init(data);
-        return result;
-    }
-
-    toJSON(data?: any) {
-        data = typeof data === 'object' ? data : {};
-        data["bookings_api_health"] = this.bookings_api_health ? this.bookings_api_health.toJSON() : <any>undefined;
-        data["user_api_health"] = this.user_api_health ? this.user_api_health.toJSON() : <any>undefined;
-        data["video_api_health"] = this.video_api_health ? this.video_api_health.toJSON() : <any>undefined;
-        return data; 
-    }
-}
-
-export interface IHealthCheckResponse {
-    bookings_api_health?: HealthCheck | undefined;
-    user_api_health?: HealthCheck | undefined;
-    video_api_health?: HealthCheck | undefined;
-}
-
 export class HealthCheck implements IHealthCheck {
-    successful?: boolean | undefined;
+    successful?: boolean;
     error_message?: string | undefined;
     data?: { [key: string]: any; } | undefined;
 
@@ -2087,15 +2237,15 @@ export class HealthCheck implements IHealthCheck {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.successful = data["successful"];
-            this.error_message = data["error_message"];
-            if (data["data"]) {
+    init(_data?: any) {
+        if (_data) {
+            this.successful = _data["successful"];
+            this.error_message = _data["error_message"];
+            if (_data["data"]) {
                 this.data = {} as any;
-                for (let key in data["data"]) {
-                    if (data["data"].hasOwnProperty(key))
-                        this.data![key] = data["data"][key];
+                for (let key in _data["data"]) {
+                    if (_data["data"].hasOwnProperty(key))
+                        this.data![key] = _data["data"][key];
                 }
             }
         }
@@ -2124,16 +2274,17 @@ export class HealthCheck implements IHealthCheck {
 }
 
 export interface IHealthCheck {
-    successful?: boolean | undefined;
+    successful?: boolean;
     error_message?: string | undefined;
     data?: { [key: string]: any; } | undefined;
 }
 
-export class AddMediaEventRequest implements IAddMediaEventRequest {
-    participant_id?: string | undefined;
-    event_type?: EventType | undefined;
+export class HealthCheckResponse implements IHealthCheckResponse {
+    bookings_api_health?: HealthCheck | undefined;
+    user_api_health?: HealthCheck | undefined;
+    video_api_health?: HealthCheck | undefined;
 
-    constructor(data?: IAddMediaEventRequest) {
+    constructor(data?: IHealthCheckResponse) {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
@@ -2142,31 +2293,34 @@ export class AddMediaEventRequest implements IAddMediaEventRequest {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.participant_id = data["participant_id"];
-            this.event_type = data["event_type"];
+    init(_data?: any) {
+        if (_data) {
+            this.bookings_api_health = _data["bookings_api_health"] ? HealthCheck.fromJS(_data["bookings_api_health"]) : <any>undefined;
+            this.user_api_health = _data["user_api_health"] ? HealthCheck.fromJS(_data["user_api_health"]) : <any>undefined;
+            this.video_api_health = _data["video_api_health"] ? HealthCheck.fromJS(_data["video_api_health"]) : <any>undefined;
         }
     }
 
-    static fromJS(data: any): AddMediaEventRequest {
+    static fromJS(data: any): HealthCheckResponse {
         data = typeof data === 'object' ? data : {};
-        let result = new AddMediaEventRequest();
+        let result = new HealthCheckResponse();
         result.init(data);
         return result;
     }
 
     toJSON(data?: any) {
         data = typeof data === 'object' ? data : {};
-        data["participant_id"] = this.participant_id;
-        data["event_type"] = this.event_type;
+        data["bookings_api_health"] = this.bookings_api_health ? this.bookings_api_health.toJSON() : <any>undefined;
+        data["user_api_health"] = this.user_api_health ? this.user_api_health.toJSON() : <any>undefined;
+        data["video_api_health"] = this.video_api_health ? this.video_api_health.toJSON() : <any>undefined;
         return data; 
     }
 }
 
-export interface IAddMediaEventRequest {
-    participant_id?: string | undefined;
-    event_type?: EventType | undefined;
+export interface IHealthCheckResponse {
+    bookings_api_health?: HealthCheck | undefined;
+    user_api_health?: HealthCheck | undefined;
+    video_api_health?: HealthCheck | undefined;
 }
 
 export enum EventType {
@@ -2187,10 +2341,58 @@ export enum EventType {
     VhoCall = "VhoCall",
 }
 
+export class AddMediaEventRequest implements IAddMediaEventRequest {
+    participant_id?: string;
+    readonly event_type?: EventType;
+
+    constructor(data?: IAddMediaEventRequest) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.participant_id = _data["participant_id"];
+            (<any>this).event_type = _data["event_type"];
+        }
+    }
+
+    static fromJS(data: any): AddMediaEventRequest {
+        data = typeof data === 'object' ? data : {};
+        let result = new AddMediaEventRequest();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["participant_id"] = this.participant_id;
+        data["event_type"] = this.event_type;
+        return data; 
+    }
+}
+
+export interface IAddMediaEventRequest {
+    participant_id?: string;
+    event_type?: EventType;
+}
+
+export enum SelfTestFailureReason {
+    Camera = "Camera",
+    Microphone = "Microphone",
+    Video = "Video",
+    BadScore = "BadScore",
+    IncompleteTest = "IncompleteTest",
+}
+
 export class AddSelfTestFailureEventRequest implements IAddSelfTestFailureEventRequest {
-    participant_id?: string | undefined;
-    event_type?: EventType | undefined;
-    self_test_failure_reason?: SelfTestFailureReason | undefined;
+    participant_id?: string;
+    readonly event_type?: EventType;
+    self_test_failure_reason?: SelfTestFailureReason;
 
     constructor(data?: IAddSelfTestFailureEventRequest) {
         if (data) {
@@ -2201,11 +2403,11 @@ export class AddSelfTestFailureEventRequest implements IAddSelfTestFailureEventR
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.participant_id = data["participant_id"];
-            this.event_type = data["event_type"];
-            this.self_test_failure_reason = data["self_test_failure_reason"];
+    init(_data?: any) {
+        if (_data) {
+            this.participant_id = _data["participant_id"];
+            (<any>this).event_type = _data["event_type"];
+            this.self_test_failure_reason = _data["self_test_failure_reason"];
         }
     }
 
@@ -2226,22 +2428,20 @@ export class AddSelfTestFailureEventRequest implements IAddSelfTestFailureEventR
 }
 
 export interface IAddSelfTestFailureEventRequest {
-    participant_id?: string | undefined;
-    event_type?: EventType | undefined;
-    self_test_failure_reason?: SelfTestFailureReason | undefined;
+    participant_id?: string;
+    event_type?: EventType;
+    self_test_failure_reason?: SelfTestFailureReason;
 }
 
-export enum SelfTestFailureReason {
-    Camera = "Camera",
-    Microphone = "Microphone",
-    Video = "Video",
-    BadScore = "BadScore",
-    IncompleteTest = "IncompleteTest",
+export enum TestScore {
+    Good = "Good",
+    Okay = "Okay",
+    Bad = "Bad",
 }
 
 export class TestCallScoreResponse implements ITestCallScoreResponse {
-    score?: TestScore | undefined;
-    passed?: boolean | undefined;
+    score?: TestScore;
+    passed?: boolean;
 
     constructor(data?: ITestCallScoreResponse) {
         if (data) {
@@ -2252,10 +2452,10 @@ export class TestCallScoreResponse implements ITestCallScoreResponse {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.score = data["score"];
-            this.passed = data["passed"];
+    init(_data?: any) {
+        if (_data) {
+            this.score = _data["score"];
+            this.passed = _data["passed"];
         }
     }
 
@@ -2275,19 +2475,13 @@ export class TestCallScoreResponse implements ITestCallScoreResponse {
 }
 
 export interface ITestCallScoreResponse {
-    score?: TestScore | undefined;
-    passed?: boolean | undefined;
-}
-
-export enum TestScore {
-    Good = "Good",
-    Okay = "Okay",
-    Bad = "Bad",
+    score?: TestScore;
+    passed?: boolean;
 }
 
 export class UpdateParticipantStatusEventRequest implements IUpdateParticipantStatusEventRequest {
-    participant_id?: string | undefined;
-    event_type?: EventType | undefined;
+    participant_id?: string;
+    readonly event_type?: EventType;
 
     constructor(data?: IUpdateParticipantStatusEventRequest) {
         if (data) {
@@ -2298,10 +2492,10 @@ export class UpdateParticipantStatusEventRequest implements IUpdateParticipantSt
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.participant_id = data["participant_id"];
-            this.event_type = data["event_type"];
+    init(_data?: any) {
+        if (_data) {
+            this.participant_id = _data["participant_id"];
+            (<any>this).event_type = _data["event_type"];
         }
     }
 
@@ -2321,12 +2515,12 @@ export class UpdateParticipantStatusEventRequest implements IUpdateParticipantSt
 }
 
 export interface IUpdateParticipantStatusEventRequest {
-    participant_id?: string | undefined;
-    event_type?: EventType | undefined;
+    participant_id?: string;
+    event_type?: EventType;
 }
 
 export class UserProfileResponse implements IUserProfileResponse {
-    role?: UserRole | undefined;
+    role?: UserRole;
     first_name?: string | undefined;
     last_name?: string | undefined;
     display_name?: string | undefined;
@@ -2340,12 +2534,12 @@ export class UserProfileResponse implements IUserProfileResponse {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.role = data["role"];
-            this.first_name = data["first_name"];
-            this.last_name = data["last_name"];
-            this.display_name = data["display_name"];
+    init(_data?: any) {
+        if (_data) {
+            this.role = _data["role"];
+            this.first_name = _data["first_name"];
+            this.last_name = _data["last_name"];
+            this.display_name = _data["display_name"];
         }
     }
 
@@ -2367,7 +2561,7 @@ export class UserProfileResponse implements IUserProfileResponse {
 }
 
 export interface IUserProfileResponse {
-    role?: UserRole | undefined;
+    role?: UserRole;
     first_name?: string | undefined;
     last_name?: string | undefined;
     display_name?: string | undefined;
@@ -2385,9 +2579,9 @@ export class SelfTestPexipResponse implements ISelfTestPexipResponse {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.pexip_self_test_node = data["pexip_self_test_node"];
+    init(_data?: any) {
+        if (_data) {
+            this.pexip_self_test_node = _data["pexip_self_test_node"];
         }
     }
 
@@ -2409,13 +2603,24 @@ export interface ISelfTestPexipResponse {
     pexip_self_test_node?: string | undefined;
 }
 
+export enum TaskType {
+    Hearing = "Hearing",
+    Judge = "Judge",
+    Participant = "Participant",
+}
+
+export enum TaskStatus {
+    ToDo = "ToDo",
+    Done = "Done",
+}
+
 export class TaskResponse implements ITaskResponse {
-    id?: number | undefined;
-    origin_id?: string | undefined;
+    id?: number;
+    origin_id?: string;
     body?: string | undefined;
-    type?: TaskType | undefined;
-    status?: TaskStatus | undefined;
-    created?: Date | undefined;
+    type?: TaskType;
+    status?: TaskStatus;
+    created?: Date;
     updated?: Date | undefined;
     updated_by?: string | undefined;
 
@@ -2428,16 +2633,16 @@ export class TaskResponse implements ITaskResponse {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.id = data["id"];
-            this.origin_id = data["origin_id"];
-            this.body = data["body"];
-            this.type = data["type"];
-            this.status = data["status"];
-            this.created = data["created"] ? new Date(data["created"].toString()) : <any>undefined;
-            this.updated = data["updated"] ? new Date(data["updated"].toString()) : <any>undefined;
-            this.updated_by = data["updated_by"];
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.origin_id = _data["origin_id"];
+            this.body = _data["body"];
+            this.type = _data["type"];
+            this.status = _data["status"];
+            this.created = _data["created"] ? new Date(_data["created"].toString()) : <any>undefined;
+            this.updated = _data["updated"] ? new Date(_data["updated"].toString()) : <any>undefined;
+            this.updated_by = _data["updated_by"];
         }
     }
 
@@ -2463,25 +2668,14 @@ export class TaskResponse implements ITaskResponse {
 }
 
 export interface ITaskResponse {
-    id?: number | undefined;
-    origin_id?: string | undefined;
+    id?: number;
+    origin_id?: string;
     body?: string | undefined;
-    type?: TaskType | undefined;
-    status?: TaskStatus | undefined;
-    created?: Date | undefined;
+    type?: TaskType;
+    status?: TaskStatus;
+    created?: Date;
     updated?: Date | undefined;
     updated_by?: string | undefined;
-}
-
-export enum TaskType {
-    Hearing = "Hearing",
-    Judge = "Judge",
-    Participant = "Participant",
-}
-
-export enum TaskStatus {
-    ToDo = "ToDo",
-    Done = "Done",
 }
 
 export class TokenResponse implements ITokenResponse {
@@ -2497,10 +2691,10 @@ export class TokenResponse implements ITokenResponse {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.expires_on = data["expires_on"];
-            this.token = data["token"];
+    init(_data?: any) {
+        if (_data) {
+            this.expires_on = _data["expires_on"];
+            this.token = _data["token"];
         }
     }
 
@@ -2524,11 +2718,51 @@ export interface ITokenResponse {
     token?: string | undefined;
 }
 
+export class HearingVenueResponse implements IHearingVenueResponse {
+    id?: number;
+    name?: string | undefined;
+
+    constructor(data?: IHearingVenueResponse) {
+        if (data) {
+            for (var property in data) {
+                if (data.hasOwnProperty(property))
+                    (<any>this)[property] = (<any>data)[property];
+            }
+        }
+    }
+
+    init(_data?: any) {
+        if (_data) {
+            this.id = _data["id"];
+            this.name = _data["name"];
+        }
+    }
+
+    static fromJS(data: any): HearingVenueResponse {
+        data = typeof data === 'object' ? data : {};
+        let result = new HearingVenueResponse();
+        result.init(data);
+        return result;
+    }
+
+    toJSON(data?: any) {
+        data = typeof data === 'object' ? data : {};
+        data["id"] = this.id;
+        data["name"] = this.name;
+        return data; 
+    }
+}
+
+export interface IHearingVenueResponse {
+    id?: number;
+    name?: string | undefined;
+}
+
 export class ConferenceEventRequest implements IConferenceEventRequest {
-    event_id?: string | undefined;
-    event_type?: EventType | undefined;
-    time_stamp_utc?: Date | undefined;
-    conference_id?: string | undefined;
+    event_id!: string | undefined;
+    event_type?: EventType;
+    time_stamp_utc?: Date;
+    conference_id!: string | undefined;
     participant_id?: string | undefined;
     transfer_from?: RoomType | undefined;
     transfer_to?: RoomType | undefined;
@@ -2543,16 +2777,16 @@ export class ConferenceEventRequest implements IConferenceEventRequest {
         }
     }
 
-    init(data?: any) {
-        if (data) {
-            this.event_id = data["event_id"];
-            this.event_type = data["event_type"];
-            this.time_stamp_utc = data["time_stamp_utc"] ? new Date(data["time_stamp_utc"].toString()) : <any>undefined;
-            this.conference_id = data["conference_id"];
-            this.participant_id = data["participant_id"];
-            this.transfer_from = data["transfer_from"];
-            this.transfer_to = data["transfer_to"];
-            this.reason = data["reason"];
+    init(_data?: any) {
+        if (_data) {
+            this.event_id = _data["event_id"];
+            this.event_type = _data["event_type"];
+            this.time_stamp_utc = _data["time_stamp_utc"] ? new Date(_data["time_stamp_utc"].toString()) : <any>undefined;
+            this.conference_id = _data["conference_id"];
+            this.participant_id = _data["participant_id"];
+            this.transfer_from = _data["transfer_from"];
+            this.transfer_to = _data["transfer_to"];
+            this.reason = _data["reason"];
         }
     }
 
@@ -2578,10 +2812,10 @@ export class ConferenceEventRequest implements IConferenceEventRequest {
 }
 
 export interface IConferenceEventRequest {
-    event_id?: string | undefined;
-    event_type?: EventType | undefined;
-    time_stamp_utc?: Date | undefined;
-    conference_id?: string | undefined;
+    event_id: string | undefined;
+    event_type?: EventType;
+    time_stamp_utc?: Date;
+    conference_id: string | undefined;
     participant_id?: string | undefined;
     transfer_from?: RoomType | undefined;
     transfer_to?: RoomType | undefined;
