@@ -42,14 +42,14 @@ namespace VideoWeb.Controllers
         {
             try
             {
-                _logger.LogTrace("Received callback from Kinly");
-                _logger.LogError($"ConferenceId: {request.Conference_id}, EventType: {request.Event_type}");
+                _logger.LogTrace("Received callback from Kinly.");
+                _logger.LogTrace($"ConferenceId: {request.Conference_id}, EventType: {request.Event_type}");
                 var callbackEvent = new CallbackEventMapper().MapConferenceEventToCallbackEventModel(request);
                 if (_memoryCache.Get<Conference>(callbackEvent.ConferenceId) == null)
                 {
                     try
                     {
-                        _logger.LogError($"Retrieving conference details for conference: {callbackEvent.ConferenceId}");
+                        _logger.LogTrace($"Retrieving conference details for conference: {callbackEvent.ConferenceId}");
                         var conference = await _videoApiClient.GetConferenceDetailsByIdAsync(callbackEvent.ConferenceId);
                         await ConferenceCache.AddConferenceToCache(conference, _memoryCache);
                     }
